refactor getTags out of createorupdatesecret
